[build] Replace random _build_id with a deterministic one
Using random values the reproducibility of iPXE, which apart from
SOURCE_DATE_EPOCH is quite good afaics. Hashing the inputs should
accomplish the same goal that the random value was used for. If
any of the inputs change we'll get a different sha1 hash which
will give a different cksum and thus a different _build_id.

If two independant binaries happen to have the same input hashes
I'd argue they are in fact the same binary.

The way I do the hashing is not quite complete, the final linked
binary could in theory give different outputs for the same input.
I don't think thats very plausible and not necessary to worry about.
This actually can happen very commonly. If building for multiple ROMs
which are built from the same blib.a, the output ends up differing due
to different arguments passed to LD. These arguments are not directly
being factored into the sha1sum. The hash/build_id is still unique
due to including version.o in the sha1 input which contains the ROM
unique ROM name.