[build] Use $(shell ...) to invoke BUILD_ID_CMD
When using $(shell ), make will first invoke run the BUILD_ID_CMD
and then have the value defined when calling $(LD). This means we
get to see the _build_id when building with make V=1. Previously
the build_id was figured out as a subshell command run during the
recipe execution with out being able to see the build_id itself.