Revert "Configure every NIC that is able to DHCP (#27)" (#29)
* Revert "Configure every NIC that is able to DHCP (#27)"

This reverts commit 5303a8bea5fb7160936043b17da2171e448eeada.

* MTL-1613 CASMINST-3789 Fix XNAME

Prior to CASMTRIAGE-2793 the netX variable was reliably relaying the
current, active DHCP interface. This behavior changed after CASMTRIAGE-2793,
where all NICs were opened instead of one-at-a-time. Thus netX can point
to any NIC that's open, and there are more NICs open at all times than the
single NIC we DHCPed with. This meant netX.dhcp/root-path sometimes
referred to our DHCPed NIC, and sometimes it did not. Thus, it would
only sometimes find the xname.

What this change does is use the index of the interface that succeeded,
instead of relying on the ambiguous netX variable we'll fetch from net4
(if 4 was the index that succeeded). This behavior is resulting in the
xname being successfully fetched.

This change removes the null setting, to prevent retries from erasing
the xname as well.