Find a segment where all modules fit
When performing a multiboot of SmartOS, it would error with
ERANGE_SEGMENT. This is due to the fact that the boot_archive was >300M,
and this wasn't fitting in the same segment the kernel occupied.

With the script:

    #!ipxe
    dhcp
    chain https://netboot.smartos.org/smartos.ipxe

The following output was seen:

    https://netboot.smartos.org/os/20220210T011229Z/platform/i86pc/kernel/amd64/unix... ok
    https://netboot.smartos.org/os/20220210T011229Z/platform/i86pc/amd64/boot_archive... ok
    https://netboot.smartos.org/os/20220210T011229Z/platform/i86pc/amd64/boot_archive.hash... ok
    Could not boot: Requested memory not available (https://ipxe.org/46038101)

This commit changes where modules are placed: after placing the image,
grab the memory map and find a segment that has space for *all* modules.
Set the module starting address to this instead of concatenating all
modules right after the image.

After this, SmartOS iPXE boots ok.